chore: turn Namespace to be a Communication
